Apply header and footer to all routes
