+++ src/components/views/Work.tsx
@@ -5,6 +5,7 @@ import { getWork } from "../../store/workApiSlice"
 import { AppDispatch } from "../../store/store"
 import { Work } from "../../types/workTypes"
 import { formatDateToMonthYear } from "../../utils/dateFormat"
+import TitleHeader from "../TitleHeader/TitleHeader"
 
 const WorkView = () => {
     const dispatch = useDispatch<AppDispatch>()
@@ -48,6 +49,7 @@ const WorkView = () => {
 
     return (
         <section>
+            <TitleHeader title="Work"/>
             <h2>Work</h2>
             {content}
         </section>
